Adding changes that were made to ToolboxLS since the last release.
These are only the files in the Examples/ and Kernel/ subdirectory (the public elements of the Subversion repository).

The view() command was removed or fixed to agree with Matlab's new interface.

The help entry for many files was modified into the new format that I'm using; however, the contents of the entry were not changed.