Adding more changes since the last release.
For some reason, not all of the new files got committed last time. I'm trying
again.

I've also modified the README file to have an introductory paragraph (since it
shows up on the BitBucket repository's home page), to refer to the bitbucket
repository's web site, and to refer to the ToolboxLS web page for the
documentation.