Fixed the remaining out-of-date view() commands in the example, plus other edits.
Searched through the Examples/ subdirectory to find the remaining examples
which contained the old version of the view command. In most cases the view
command was deleted because the example was in 2D so the view did not matter.

Also added '\n' before and after the final computation time fprintf() in those
same files (didn't bother to look in the other files).

Tried to remove as many m-lint warnings as possible from those same files
(didn't bother to look in the other files).

Finally, I deleted all the .asv and *~ (eg: backup) files that I found,
although that shouldn't be visible in the repository.